requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
DateTime==5.5
Flask==3.0.3
fonttools==4.54.1
gcloud==0.18.3
geopandas==1.0.1
googleapis-common-protos==1.65.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.1
matplotlib==3.9.2
numpy==2.1.2
oauth2client==4.1.3
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==5.29.0rc1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pygrib==2.1.6
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rsa==4.9
schedule==1.2.2
setuptools==75.1.0
shapely==2.0.6
six==1.16.0
soupsieve==2.6
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
zope.interface==7.1.0
